Adjust sockets for handle based version
Use chromium socket API based on handles not objects.
For syscall 102 (typical socket functions) disable proxy to render
thread. For syscalls 3, 4 and 6 (read, write, close) disable proxy if
argument fd is a socket descriptor. For syscall 142 and 168 (select, poll)
find first fd and check if it is a socket descriptor. For socket descriptor
disable proxy.

Signed-off-by: Grzegorz Wolny <g.wolny@samsung.com>
Change-Id: I4b8a639534f45f2f9f391375c4d84cbba7f7863b