Filesystem support in JS forced enabled
As it is stated in src/settings.js, the flag FORCE_FILESYSTEM might be needed
in cases when filesystem is used in JS and that might not be detected properly.
As it turns out this case occurs in our situation (isSocket() function not defined,
despite the fact that it is used).

Change-Id: Ib3ed996d18b3d1c30e35675e7ba694824ee57147
Signed-off-by: Robert Bycul <r.bycul@samsung.com>
Bug: https://cam.sprc.samsung.pl/browse/VDWASM-277