[EMSS] Modified setting framerate and resolution
Changed a way that framerate and resolution is set in glue code.
Until now, when one of resolution (width, height) or framerate (num, den)
parameters was 0, other one was set as 0 too - despite of its actual
value.

It did not return any error and masked a problem that one of parameters
is missing - there is already check for such situation in chromium,
which until now was not triggered (when both parameters are set to 0
we assume that resolution/framerate is not set).

This change sets all mentioned parameters with their exact values.
All checks regarding framerate and resolution will be performed in
chromium.

Bug: https://cam.sprc.samsung.pl/browse/VDWASM-315
Signed-off-by: Szczepan Radomski <s.radomski@samsung.com>
Change-Id: I43c35a2bc24b338ea82db305a15d55caf77ce66e