Allow pthread_mutex_timedlock to be cancelled properly
When pthread_mutex_timedlock was blocked, waiting for
another thread to release mutex and thread was cancelled
asynchronously, pthread_mutex_timedlock returned with
errno 11 (ECANCELED).
This is not compatible with POSIX standard. This patch
fixes that behaviour, pthread is properly canceled in
such situation.
Special check at the end of pthread_mutex_timedlock has
been added.

Bug: https://cam.sprc.samsung.pl/browse/VDWASM-557
Change-Id: Ic65de29ba4aa0b066479f9c8df5feab94ed8a647