[WASM Player] Add checking sample format to EMSS adding audio track
Current WASM Player implementations support PCM in S16LE format only.
This patch adds an early validation of EMSS audio track config.
If audio codec is PCM and sample format is different than S16,
an error is returned to the user.

Proper validation is done in WASM player implementation.
Old versions of platform API didn't perform PCM format validation,
and this is a workaround for that - if someone uses a newer
toolchain, they will get an error from here.

Bug: https://cam.sprc.samsung.pl/browse/VDWASM-626
Signed-off-by: Kamil Kaznowski <k.kaznowski@samsung.com>
Change-Id: I187eadc07ee02bedc6851f63b5efa9b5ffcbacbf